move utils unit tests to separate file so it's easy to find accompanying tests
